+++ .github/workflows/test.yaml
@@ -2,9 +2,9 @@ name: Run tests
 
 on:
   push:
-    branches: [ "master" ]
+    branches: [ "main", "dev", "stage" ]
   pull_request:
-    branches: [ "master" ]
+    branches: [ "main", "dev", "stage" ]
 
 env:
   CARGO_TERM_COLOR: always
