+++ .gitignore
@@ -1,27 +1,32 @@
-# Runtime/test state
-/datacore/
-/dataindex/
-/dataldk0/
-/dataldk1/
-/dataldk2/
-/tmp/
+# Test files
+/datacore
+/dataindex
+/dataldk0
+/dataldk1
+/dataldk2
+/tmp
 
-# Rust build outputs
-/target/
-/bindings/uniffi-bindgen/target/
+# will have compiled files and executables
+/target
+
+# These are backup files generated by rustfmt
 **/*.rs.bk
 
-# IDE/editor/OS files
+# intellij files
 .idea/
 .vscode/
+
+# MacOS files
 .DS_Store
 *.swp
 
-# Generated UniFFI artifacts (regenerated in CI/local scripts)
-/target/uniffi/
-/bindings/swift/
-/bindings/kotlin/
-/bindings/kotlin-android/org/
-/bindings/kotlin-android/jniLibs/
-/bindings/python/dist/
+/bindings/uniffi-bindgen/target
+/target/uniffi
+/bindings/swift
+/bindings/kotlin
+/bindings/kotlin-android/org
+/bindings/kotlin-android/jniLibs
+/bindings/python/dist
 /bindings/python/rgb_lightning_node.py
+/artifacts
+/bindings/python/build
