+++ Cargo.toml
@@ -86,22 +86,7 @@ level = "allow"
 inherits = "release"
 debug = 1
 
-[[test]]
-name = "lib_core_uniffi_state"
-path = "tests/lib_core/uniffi_state.rs"
-required-features = ["uniffi", "test-utils"]
-
-[[test]]
-name = "lib_core_type_converters"
-path = "tests/lib_core/type_converters.rs"
-required-features = ["uniffi", "test-utils"]
-
-[[test]]
-name = "lib_core_error_mapping"
-path = "tests/lib_core/error_mapping.rs"
-required-features = ["uniffi", "test-utils"]
-
-[[test]]
-name = "lib_core_node_handle_invariants"
-path = "tests/lib_core/node_handle_invariants.rs"
+[[est]]
+name = "lib_core"
+path = "tests/lib_core/lib_core.rs"
 required-features = ["uniffi", "test-utils"]
